vertex program skeleton working for degrees
