+++ mindmaps-test/src/test/java/io/mindmaps/graql/internal/analytics/AnalyticsTest.java
@@ -258,7 +258,7 @@ public class AnalyticsTest {
         });
     }
 
-    @Ignore
+//    @Ignore
     @Test
     public void testDegreesAndPersist() throws Exception {
         // create instances
